feat(navbar): Add icon props to NavItem
This commit adds a new `icon` prop to the `INavItem` interface in `nav.d.ts`. It also updates the `NavBarButton` component to use the `icon` prop instead of the `src` prop. The `NavBarButton` component now renders an `Image` component with the `src`, `width`, `height`, and `alt` props from the `icon` prop. The `subNavItemsConfig` in `navBarConfig.ts` is also updated to use the new `icon` prop instead of the `src` prop for the 'General Events' NavItem.